pyproject.toml
[tool.poetry]
name = "task-gku-sar"
version = "0.1.0"
description = ""
authors = ["Anastasia Maximova <aastavitskaya@gmail.com>", "Roman Ryzhkov <romka_ne@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
openpyxl = "^3.1.2"
aiohttp = "^3.9.0"
isort = "^5.12.0"
black = "^23.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
